feat: store paginated collections as a dict
This ensures that the collection (like issues and PRs) are unique since
keys can't be duplicated in a dict. This is also better for checking if
an item exists than looping over lists.

Note that when items share the same key, the oldest one is kept since it
overwrites the dict's value.

Aside from this, we're also removing PRs from the issues list since not
only is this redundant, but they can't be linked to or from sub-issues,
which can cause conflicts.